move: cache package와 lock package를 분리
