+++ src/main/java/com/newzet/api/common/lock/exception/LockAcquisitionException.java
@@ -1,4 +1,4 @@
-package com.newzet.api.common.cache.exception;
+package com.newzet.api.common.lock.exception;
 
 public class LockAcquisitionException extends RuntimeException {
 	public LockAcquisitionException(String route) {
