+++ src/main/java/com/newzet/api/newsletter/business/NewsletterService.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.newzet.api.common.cache.CacheUtil;
-import com.newzet.api.common.cache.LockFactory;
-import com.newzet.api.common.cache.exception.LockAcquisitionException;
+import com.newzet.api.common.lock.LockFactory;
 import com.newzet.api.newsletter.business.dto.NewsletterCacheDto;
 import com.newzet.api.newsletter.business.dto.NewsletterEntityDto;
 import com.newzet.api.newsletter.domain.Newsletter;
