+++ src/test/java/com/newzet/api/newsletter/business/NewsletterServiceTest.java
@@ -13,8 +13,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.newzet.api.common.cache.CacheUtil;
-import com.newzet.api.common.cache.LockFactory;
-import com.newzet.api.common.cache.exception.LockAcquisitionException;
+import com.newzet.api.common.lock.LockFactory;
 import com.newzet.api.newsletter.business.dto.NewsletterCacheDto;
 import com.newzet.api.newsletter.business.dto.NewsletterEntityDto;
 import com.newzet.api.newsletter.domain.Newsletter;
